Reorganize files to enable dynamic testing
