+++ packages/parser/src/componentBuilders/fragments/parsingClasses/Accordion.ts
@@ -1,6 +1,6 @@
-import { FragmentViaBxmlTag } from "./FragmentViaBxmlTag";
+import { FragmentViaBxmlTag } from "../abstractClasses/FragmentViaBxmlTag";
 import { z } from "zod";
-import { FragmentViaBxmlTagParams } from "../../types/fragments";
+import { FragmentViaBxmlTagParams } from "../../../types/fragments";
 import { AccordionItemTagSchema, AccordionItemTag } from "./AccordionItem";
 import { RawFragmentSchema } from "@b5x/types";
