+++ packages/parser/src/componentBuilders/fragments/parsingClasses/SelectQuestion.ts
@@ -1,7 +1,7 @@
-import { BxmlTextNodeSchema } from "../../types/bxmlNodes";
-import { FragmentViaBxmlTag } from "./FragmentViaBxmlTag";
+import { BxmlTextNodeSchema } from "../../../types/bxmlNodes";
+import { FragmentViaBxmlTag } from "../abstractClasses/FragmentViaBxmlTag";
 import { z } from "zod";
-import { FragmentViaBxmlTagParams } from "../../types/fragments";
+import { FragmentViaBxmlTagParams } from "../../../types/fragments";
 
 // Markup -----------------------------------------------------------
