+++ packages/parser/src/tests/fragments/parsingClasses.test.js
@@ -1,0 +1,5 @@
+describe("placeholder", () => { 
+  test("placeholder", () => {
+    expect(true).toBe(true);
+  });
+});
