+++ packages/parser/src/typedocEntryPoint.ts
@@ -2,36 +2,36 @@ import { BoomboxParser } from "./index";
 import { Topic } from "./componentBuilders/Topic";
 import { Document } from "./componentBuilders/Document";
 import { FragmentBuilder } from "./componentBuilders/FragmentBuilder";
-import { Fragment } from "./componentBuilders/fragments/Fragment";
-import { Accordion } from "./componentBuilders/fragments/Accordion";
-import { AccordionItem } from "./componentBuilders/fragments/AccordionItem";
-import { Badge } from "./componentBuilders/fragments/Badge";
-import { Breakout } from "./componentBuilders/fragments/Breakout";
-import { Checklist } from "./componentBuilders/fragments/Checklist";
-import { CodeBlock } from "./componentBuilders/fragments/CodeBlock";
-import { Diagram } from "./componentBuilders/fragments/Diagram";
-import { Echo } from "./componentBuilders/fragments/Echo";
-import { FiveStarRating } from "./componentBuilders/fragments/FiveStarRating";
-import { FragmentViaArgs } from "./componentBuilders/fragments/FragmentViaArgs";
-import { FragmentViaBxmlTag } from "./componentBuilders/fragments/FragmentViaBxmlTag";
-import { FragmentViaBxmlText } from "./componentBuilders/fragments/FragmentViaBxmlText";
-import { LongTextQuestion } from "./componentBuilders/fragments/LongTextQuestion";
-import { MermaidDiagram } from "./componentBuilders/fragments/MermaidDiagram";
-import { NpsQuestion } from "./componentBuilders/fragments/NpsQuestion";
-import { Rubric } from "./componentBuilders/fragments/Rubric";
-import { SelectQuestion } from "./componentBuilders/fragments/SelectQuestion";
-import { SentimentCheck } from "./componentBuilders/fragments/SentimentCheck";
-import { ShortTextQuestion } from "./componentBuilders/fragments/ShortTextQuestion";
-import { Show } from "./componentBuilders/fragments/Show";
-import { Slide } from "./componentBuilders/fragments/Slide";
-import { Slideshow } from "./componentBuilders/fragments/Slideshow";
-import { Step } from "./componentBuilders/fragments/Step";
-import { Tab } from "./componentBuilders/fragments/Tab";
-import { Tabs } from "./componentBuilders/fragments/Tabs";
-import { Troubleshooter } from "./componentBuilders/fragments/Troubleshooter";
-import { ContinueButton } from "./componentBuilders/fragments/ContinueButton";
-import { VisualList } from "./componentBuilders/fragments/VisualList";
-import { Warning } from "./componentBuilders/fragments/Warning";
+import { Fragment } from "./componentBuilders/fragments/abstractClasses/Fragment";
+import { Accordion } from "./componentBuilders/fragments/parsingClasses/Accordion";
+import { AccordionItem } from "./componentBuilders/fragments/parsingClasses/AccordionItem";
+import { Badge } from "./componentBuilders/fragments/parsingClasses/Badge";
+import { Breakout } from "./componentBuilders/fragments/parsingClasses/Breakout";
+import { Checklist } from "./componentBuilders/fragments/parsingClasses/Checklist";
+import { CodeBlock } from "./componentBuilders/fragments/parsingClasses/CodeBlock";
+import { Diagram } from "./componentBuilders/fragments/parsingClasses/Diagram";
+import { Echo } from "./componentBuilders/fragments/parsingClasses/Echo";
+import { FiveStarRating } from "./componentBuilders/fragments/parsingClasses/FiveStarRating";
+import { FragmentViaArgs } from "./componentBuilders/fragments/parsingClasses/FragmentViaArgs";
+import { FragmentViaBxmlTag } from "./componentBuilders/fragments/abstractClasses/FragmentViaBxmlTag";
+import { FragmentViaBxmlText } from "./componentBuilders/fragments/abstractClasses/FragmentViaBxmlText";
+import { LongTextQuestion } from "./componentBuilders/fragments/parsingClasses/LongTextQuestion";
+import { MermaidDiagram } from "./componentBuilders/fragments/parsingClasses/MermaidDiagram";
+import { NpsQuestion } from "./componentBuilders/fragments/parsingClasses/NpsQuestion";
+import { Rubric } from "./componentBuilders/fragments/parsingClasses/Rubric";
+import { SelectQuestion } from "./componentBuilders/fragments/parsingClasses/SelectQuestion";
+import { SentimentCheck } from "./componentBuilders/fragments/parsingClasses/SentimentCheck";
+import { ShortTextQuestion } from "./componentBuilders/fragments/parsingClasses/ShortTextQuestion";
+import { Show } from "./componentBuilders/fragments/parsingClasses/Show";
+import { Slide } from "./componentBuilders/fragments/parsingClasses/Slide";
+import { Slideshow } from "./componentBuilders/fragments/parsingClasses/Slideshow";
+import { Step } from "./componentBuilders/fragments/parsingClasses/Step";
+import { Tab } from "./componentBuilders/fragments/parsingClasses/Tab";
+import { Tabs } from "./componentBuilders/fragments/parsingClasses/Tabs";
+import { Troubleshooter } from "./componentBuilders/fragments/parsingClasses/Troubleshooter";
+import { ContinueButton } from "./componentBuilders/fragments/parsingClasses/ContinueButton";
+import { VisualList } from "./componentBuilders/fragments/parsingClasses/VisualList";
+import { Warning } from "./componentBuilders/fragments/parsingClasses/Warning";
 
 export {
   BoomboxParser,
